requirements.txt
numpy
tf_keras
tensorflow==2.13.0
pandas
pyarrow==15
notebook
matplotlib
seaborn
scipy
joblib
python-box==5.4.1
pyyaml
types-PyYAML
tqdm
ensure==1.0.2
ipykernel
streamlit
pillow
nltk
gdown
contractions
emoji
scikit-learn
tensorflow_hub
tensorflow_text

-e .


